RST-4472 Added JSON server to mimick the new JSON service from acas
