+++ lib/et_fake_acas_server/version.rb
@@ -1,3 +1,3 @@
 module EtFakeAcasServer
-  VERSION = "0.1.12"
+  VERSION = "0.2.0"
 end
